Refactor: 재사용 가능한 PageInfo 구현
