+++ src/app/(afterLogin)/admin/content/badge/page.tsx
@@ -7,8 +7,8 @@ import {
 import Layout from '@/share/layout/Layout';
 import { usePrefetchBadgeList } from '@/share/query/badge/useGetBadgeList';
 
+import PageInfo from '../../_components/PageInfo';
 import BadgeList from './_components/BadgeList';
-import BadgePageInfo from './_components/BadgePageInfo';
 import CreateBadge from './_components/CreateBadge';
 
 export default async function Badge() {
@@ -17,7 +17,7 @@ export default async function Badge() {
   return (
     <>
       <Layout.Content.Full>
-        <BadgePageInfo />
+        <PageInfo title="배지 관리" path={['콘텐츠', '배지 관리']} />
       </Layout.Content.Full>
       <HydrationBoundary state={dehydrate(queryClient)}>
         <Layout.Content.Full>
