+++ src/app/(afterLogin)/admin/content/rollingBanner/[bannerId]/page.tsx
@@ -9,8 +9,8 @@ import { usePrefetchBanner } from '@/share/query/banner/useGetBanner';
 import { usePrefetchBannerList } from '@/share/query/banner/useGetBannerList';
 import { usePrefetchVocaList } from '@/share/query/voca/useGetVocaList';
 
+import PageInfo from '../../../_components/PageInfo';
 import BannerList from '../_components/BannerList';
-import BannerPageInfo from '../_components/BannerPageInfo';
 import UpdateBanner from '../_components/UpdateBanner';
 
 export default async function Banner({
@@ -33,7 +33,7 @@ export default async function Banner({
   return (
     <>
       <Layout.Content.Full>
-        <BannerPageInfo />
+        <PageInfo title="롤링 배너 관리" path={['콘텐츠', '롤링 배너 관리']} />
       </Layout.Content.Full>
       <HydrationBoundary state={dehydrate(queryClient)}>
         <Layout.Content.Full>
