+++ src/app/(afterLogin)/admin/content/rollingBanner/page.tsx
@@ -8,8 +8,8 @@ import Layout from '@/share/layout/Layout';
 import { usePrefetchBannerList } from '@/share/query/banner/useGetBannerList';
 import { usePrefetchVocaList } from '@/share/query/voca/useGetVocaList';
 
+import PageInfo from '../../_components/PageInfo';
 import BannerList from './_components/BannerList';
-import BannerPageInfo from './_components/BannerPageInfo';
 import CreateBanner from './_components/CreateBanner';
 
 export default async function Banner() {
@@ -23,7 +23,7 @@ export default async function Banner() {
   return (
     <>
       <Layout.Content.Full>
-        <BannerPageInfo />
+        <PageInfo title="롤링 배너 관리" path={['콘텐츠', '롤링 배너 관리']} />
       </Layout.Content.Full>
       <HydrationBoundary state={dehydrate(queryClient)}>
         <Layout.Content.Full>
