+++ src/app/(afterLogin)/admin/content/themeCard/[cardId]/page.tsx
@@ -10,8 +10,8 @@ import { usePrefetchCardList } from '@/share/query/card/useGetCardList';
 import { usePrefetchQuizList } from '@/share/query/quiz/useGetQuizList';
 import { usePrefetchVocaList } from '@/share/query/voca/useGetVocaList';
 
+import PageInfo from '../../../_components/PageInfo';
 import CardList from '../_components/CardList';
-import CardPageInfo from '../_components/CardPageInfo';
 import UpdateCard from '../_components/UpdateCard';
 
 export default async function Card({ params }: { params: { cardId: string } }) {
@@ -33,7 +33,10 @@ export default async function Card({ params }: { params: { cardId: string } }) {
   return (
     <>
       <Layout.Content.Full>
-        <CardPageInfo />
+        <PageInfo
+          title="테마별 카드모음집 관리"
+          path={['콘텐츠', '테마별 카드모음집 관리']}
+        />
       </Layout.Content.Full>
       <HydrationBoundary state={dehydrate(queryClient)}>
         <Layout.Content.Full>
