+++ src/app/(afterLogin)/admin/quiz/vocaManagement/page.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 
 import CardFilter from '@/app/(afterLogin)/admin/quiz/_components/CardFilter';
-import PageInfo from '@/app/(afterLogin)/admin/quiz/_components/PageInfo';
 import CreateVoca from '@/app/(afterLogin)/admin/quiz/vocaManagement/_components/CreateVoca';
 import VocaSearchList from '@/app/(afterLogin)/admin/quiz/vocaManagement/_components/VocaSearchList';
 import Layout from '@/share/layout/Layout';
+
+import PageInfo from '../../_components/PageInfo';
 
 export default async function TermManagement() {
   return (
